[SPARK-XXXXX][SQL][TESTS] Revert SQLTestUtils merge into QueryTest
SQLTestUtils cannot extend QueryTest (abstract class) because many
test suites mix SQLTestUtils with SparkFunSuite or SparkPlanTest
directly. Methods that call test() must stay in a class/trait that
extends SparkFunSuite. Reverting to keep SQLTestUtils unchanged.

This PR now only merges SQLTestUtilsBase into QueryTestBase.
